refactor(request): use rustic-error and improve architecture
