+++ src/api/requests/main-channel.ts
@@ -1,4 +1,4 @@
-import { graphql } from "$core/utils/request/graphql";
+import { graphql } from "$core/utils/request/graphql/code-gen";
 
 export const addChannel = graphql(`
     mutation addChannel($channelId: ID!, $category: String!){
