+++ src/commands/anniversaire/[sub-commands]/next.cmd.ts
@@ -2,7 +2,7 @@ import type { CommandExecute } from "$core/utils/handler/command";
 import { getBirthdays } from "$core/api/requests/member";
 import { simpleEmbed } from "$core/utils/embed";
 import { commands } from "$core/configs/message/command";
-import { gqlRequest } from "$core/utils/request";
+import { gqlRequest } from "$core/utils/request/graphql/code-gen";
 import { DayJS } from "$core/utils/day-js";
 import { msgParams } from "$core/utils/message";
 import { getAge } from "$core/utils/function";
