+++ src/commands/main/[sub-commands]/group-add/add-role.cmd.ts
@@ -2,7 +2,7 @@ import type { CommandExecute } from "$core/utils/handler/command";
 import { addRole } from "$core/api/requests/main-role";
 import { simpleEmbed } from "$core/utils/embed";
 import { commands } from "$core/configs/message/command";
-import { gqlRequest } from "$core/utils/request";
+import { gqlRequest } from "$core/utils/request/graphql/code-gen";
 import { msgParams } from "$core/utils/message";
 import { logger } from "$core/utils/logger";
 import { userWithId } from "$core/utils/user";
