+++ src/commands/main/[sub-commands]/group-remove/remove-channel.cmd.ts
@@ -2,7 +2,7 @@ import { removeChannel } from "$core/api/requests/main-channel";
 import { simpleEmbed } from "$core/utils/embed";
 import { msgParams } from "$core/utils/message";
 import type { CommandExecute } from "$core/utils/handler/command";
-import { gqlRequest } from "$core/utils/request";
+import { gqlRequest } from "$core/utils/request/graphql/code-gen";
 import { commands } from "$core/configs/message/command";
 import { logger } from "$core/utils/logger";
 import { userWithId } from "$core/utils/user";
