+++ src/commands/stats/stats.type.ts
@@ -1,4 +1,4 @@
-import type { ServerActivity } from "$core/utils/request/graphql/graphql";
+import type { ServerActivity } from "$core/utils/request/graphql/code-gen/graphql";
 
 export type GraphType = {
   columnName: keyof Omit<ServerActivity, "date" | "__typename">;
