+++ src/commands/top-message/[sub-commands]/channel.cmd.ts
@@ -6,7 +6,7 @@ import { formatPage, getPage } from "../top-message.util";
 import { commands } from "$core/configs/message/command";
 import { simpleEmbed } from "$core/utils/embed";
 import { msgParams } from "$core/utils/message";
-import { gqlRequest } from "$core/utils/request";
+import { gqlRequest } from "$core/utils/request/graphql/code-gen";
 import { logger } from "$core/utils/logger";
 import { userWithId } from "$core/utils/user";
