+++ src/events/github-link-reaction/github-link-reaction.event.ts
@@ -2,7 +2,7 @@ import type { CodeElement } from "./github-link-reaction.type";
 import type { EventExecute, EventName } from "$core/utils/handler/event";
 import { githubPermalinkRegex, githubRawUrl } from "./github-link-reaction.const";
 import { getStringEnv } from "$core/utils/env-variable";
-import { restTextRequest } from "$core/utils/request/request";
+import { restTextRequest } from "$core/utils/request";
 import { logger } from "$core/utils/logger";
 import { userWithId } from "$core/utils/user";
