+++ src/events/guild-member-add/guild-member-add.event.ts
@@ -6,7 +6,7 @@ import { defaultAvatar } from "./guild-member-add.const";
 import { simpleEmbed } from "$core/utils/embed";
 import { logger } from "$core/utils/logger";
 import { msgParams } from "$core/utils/message";
-import { gqlRequest } from "$core/utils/request";
+import { gqlRequest } from "$core/utils/request/graphql/code-gen";
 import { ChannelType } from "discord.js";
 import { userWithId } from "$core/utils/user";
