+++ src/events/user-update/user-update.event.ts
@@ -1,5 +1,5 @@
 import type { EventExecute, EventName } from "$core/utils/handler/event";
-import { gqlRequest } from "$core/utils/request";
+import { gqlRequest } from "$core/utils/request/graphql/code-gen";
 import { setUsernameAndProfilePicture } from "$core/api/requests/member";
 import { logger } from "$core/utils/logger";
 import { userWithId } from "$core/utils/user";
