+++ src/tasks/tier-update-game/tier-update-game.task.ts
@@ -7,7 +7,7 @@ import { simpleEmbed } from "$core/utils/embed";
 import type { TaskExecute, TaskInterval } from "$core/utils/handler/task";
 import { logger } from "$core/utils/logger";
 import { msgParams } from "$core/utils/message";
-import { gqlRequest } from "$core/utils/request";
+import { gqlRequest } from "$core/utils/request/graphql/code-gen";
 import { BaseGuildTextChannel } from "discord.js";
 import { userWithId } from "$core/utils/user";
 import { getGuildMembers } from "$core/utils/discord/guild";
