+++ src/utils/request/graphql/code-gen/index.ts
@@ -1,0 +1,2 @@
+export * from "./fragment-masking";
+export * from "./gql";
