+++ src/utils/request/graphql/index.ts
@@ -1,2 +1,2 @@
-export * from "./fragment-masking";
-export * from "./gql";
+export * from "./graphql"
+export * from "./code-gen"
