--- src/utils/request/request.type.ts
@@ -1,10 +1,0 @@
-export type ResponseSuccess<T> = {
-  success: true;
-  data: T;
-}
-
-export type ResponseError = {
-  success: false;
-}
-
-export type Response<T> = ResponseSuccess<T> | ResponseError;
