+++ src/utils/request/rest/index.ts
@@ -1,0 +1,2 @@
+export * from "./rest.util";
+export * from "./rest.type";
