+++ src/utils/request/rest/rest.type.ts
@@ -1,0 +1,5 @@
+export type Method = "get" | "delete" | "post"| "put";
+
+export type RequestParams = Omit<RequestInit, "method"> & {
+  query?: Record<string, string | string[]>;
+}
